refactor: streamline species configuration handling and improve action processing
- Removed the deprecated species configuration loading logic, consolidating species settings into a new structure for better clarity and maintainability.
- Updated action processing to utilize the new species thresholds and action configurations, enhancing the filtering logic for species inclusion/exclusion.
- Introduced helper functions for better management of script parameters and action handling, ensuring a more modular approach.
- Improved logging for debugging purposes, providing clearer insights into species configuration usage during processing.
